Quiz2Screen 화면이동 함수 연결

+++ app/src/main/java/com/sandy/memorizingvoca/ui/feature/voca_list/VocaListScreen.kt
@@ -18,7 +18,7 @@ internal fun VocaListRoute(
     onNavigateBack: () -> Unit,
     onNavigateFull: () -> Unit,
     onNavigateQuiz1: (Int) -> Unit,
-    onNavigateQuiz2: () -> Unit,
+    onNavigateQuiz2: (Int) -> Unit,
     onItemClick: (Int) -> Unit,
     viewModel: VocaListViewModel = hiltViewModel(),
 ) {
@@ -41,7 +41,7 @@ private fun VocaListScreen(
     onNavigateBack: () -> Unit,
     onNavigateFull: () -> Unit,
     onNavigateQuiz1: (Int) -> Unit,
-    onNavigateQuiz2: () -> Unit,
+    onNavigateQuiz2: (Int) -> Unit,
     onItemClick: (Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
